+++ src/pages/ObservationPage/ObservationPage.tsx
@@ -36,6 +36,7 @@ export default function ObservationPage() {
   const [elementsO, setElementsO] = React.useState(null);
   const [elementsT, setElementsT] = React.useState(null);
   const [row, setRow] = React.useState(null);
+  const [obRow, setObRow] = React.useState(0);
 
   const getProposal = () => application.content2 as Proposal;
   const setProposal = (proposal: Proposal) => updateAppContent2(proposal);
@@ -437,7 +438,11 @@ export default function ObservationPage() {
                 columns={extendedColumnsObservations}
                 height={450}
                 onRowClick={e => setCurrObs(e.row.rec)}
+                onRowSelectionModelChange={e => {
+                  setObRow(e);
+                }}
                 testId="observationDetails"
+                rowSelectionModel={obRow}
               />
             )}
             {!hasObservations() && (